backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.1
notion-client==2.2.1
openai==1.3.9
apscheduler==3.10.4
requests==2.31.0
